requirements.txt
kafka-python==2.0.2
cryptography==41.0.7
passlib==1.7.4
pywin32==306; sys_platform == 'win32'
wmi==1.5.1; sys_platform == 'win32'
